fix(connections): load full connection lists
